Move `RootCertStore` to tls/ext
