+++ src/tls/mod.rs
@@ -9,18 +9,19 @@ mod conn;
 mod ext;
 
 use crate::impl_debug;
+use boring2::ssl::SslCurve;
 use boring2::{
     error::ErrorStack,
     ssl::{SslConnector, SslMethod, SslOptions, SslVersion},
 };
-use boring2::{ssl::SslCurve, x509::store::X509Store};
 use conn::{HttpsLayer, HttpsLayerSettings};
 use std::borrow::Cow;
 use typed_builder::TypedBuilder;
 
 pub use conn::{HttpsConnector, MaybeHttpsStream};
 pub use ext::{
-    CertCompressionAlgorithm, ConnectConfigurationExt, SslConnectorBuilderExt, SslRefExt,
+    CertCompressionAlgorithm, ConnectConfigurationExt, RootCertStore, SslConnectorBuilderExt,
+    SslRefExt,
 };
 
 type TlsResult<T> = Result<T, ErrorStack>;
@@ -190,63 +191,6 @@ impl TlsInfo {
     /// Get the DER encoded leaf certificate of the peer.
     pub fn peer_certificate(&self) -> Option<&[u8]> {
         self.peer_certificate.as_ref().map(|der| &der[..])
-    }
-}
-
-/// The root certificate store.
-#[allow(missing_debug_implementations)]
-#[derive(Default)]
-pub enum RootCertStore {
-    /// An owned `X509Store`.
-    Owned(X509Store),
-
-    /// A borrowed `X509Store`.
-    Borrowed(&'static X509Store),
-
-    /// Use the system's native certificate store.
-    #[default]
-    Default,
-}
-
-/// ====== impl RootCertsStore ======
-macro_rules! impl_root_cert_store {
-    ($($type:ty => $variant:ident),* $(,)?) => {
-        $(
-            impl From<$type> for RootCertStore {
-                fn from(store: $type) -> Self {
-                    Self::$variant(store)
-                }
-            }
-        )*
-    };
-
-    ($($type:ty => $variant:ident, $unwrap:expr),* $(,)?) => {
-        $(
-            impl From<$type> for RootCertStore {
-                fn from(store: $type) -> Self {
-                    $unwrap(store).map(Self::$variant).unwrap_or_default()
-                }
-            }
-        )*
-    };
-}
-
-impl_root_cert_store!(
-    X509Store => Owned,
-    &'static X509Store => Borrowed,
-);
-
-impl_root_cert_store!(
-    Option<X509Store> => Owned, |s| s,
-    Option<&'static X509Store> => Borrowed, |s| s,
-);
-
-impl<F> From<F> for RootCertStore
-where
-    F: Fn() -> Option<&'static X509Store>,
-{
-    fn from(func: F) -> Self {
-        func().map(Self::Borrowed).unwrap_or_default()
     }
 }
